Added interface for ExcelProcess.java
